projects/requirements.txt
appdirs==1.4.4
asgiref==3.3.4
autopep8==1.5.6
beautifulsoup4==4.9.3
black==21.4b1
boto3==1.17.95
botocore==1.20.95
CacheControl==0.12.10
cachetools==5.0.0
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
choices==0.1
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
decorator==5.1.0
deep-translator==1.4.4
Django==3.2
django-cors-headers==3.7.0
django-debug-toolbar==3.2.1
django-environ==0.4.5
django-extensions==3.1.3
django-filter==2.4.0
django-model-utils==4.1.1
django-mysql==3.11.1
django-query-logger==0.1.2
django-rql==3.9.0
django-simple-history==3.1.1
django-sql-sniffer==1.0.3
django-treebeard==4.5.1
django-werkzeug==1.0.0
djangorestframework==3.12.4
djangorestframework-datatables==0.6.0
djangorestframework-simplejwt==5.1.0
drf-yasg==1.20.0
fcm-django==1.0.11
firebase-admin==5.2.0
google-api-core==2.7.1
google-api-python-client==2.42.0
google-auth==2.6.2
google-auth-httplib2==0.1.0
google-cloud-core==2.2.3
google-cloud-firestore==2.4.0
google-cloud-storage==2.2.1
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
graphviz==0.17
grpcio==1.45.0
grpcio-status==1.45.0
httplib2==0.20.4
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
jmespath==0.10.0
jsonfield==3.1.0
lark-parser==0.8.2
MarkupSafe==1.1.1
msgpack==1.0.3
mypy-extensions==0.4.3
mysqlclient==2.0.3
openapi-codec==1.3.2
packaging==20.9
pathspec==0.8.1
Pillow==8.2.0
proto-plus==1.20.3
protobuf==3.20.0
py-firebase-dynamic-links==1.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.21
pydot==1.4.2
pyfcm==1.5.4
Pygments==2.9.0
PyJWT==2.4.0
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
regex==2021.4.4
requests==2.25.1
required==0.4.0
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.4.2
self==2020.12.3
simplejson==3.17.2
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
stripe==2.72.0
toml==0.10.2
twilio==7.1.0
typed-ast==1.4.3
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
validators==0.19.0
Werkzeug==2.0.1